fix workflow file from github actions
When is created a workflow file in the GitHub actions, they not update this file anymore, this way when I created the file on pr #5 this wad a typo in the `releases` and broken the job. I fix this file in the pr, but GitHub not updated this file :(

I think this is a bug from GitHub actions and already an issue in the GitHub community https://github.community/t5/GitHub-Actions/Workflow-file-not-updating-anymore

To fix the problem I renamed the file and created new commit.