add konduto prefix in all class
To avoid ambiguous names when someone uses in your domain, for example, whether someone uses this SDK in e-commerce in the code is very possible that have:

Customer -> Domain
Customer -> Konduto

and when we read the code is very complex to understand the difference between the customer class.

with these changes the code is more legible:

Customer -> Domain
KondutoCustomer -> Konduto.