remove unnecessary classes and add missing order status
Some classes were unnecessary to make the payload to send to konduto, this way I removed from the project.

Another change is in some corner cases of case sensitive to create KondutoOrderStatus or KondutoRecommentation based on the response from Konduto.

For the last, I added some missing KondutoOrderStatus like PENDING and NOT_ANALYZED when some order return this status broken the app because she can't parse the value